requirements.txt
cryptography==3.1.1
seaborn==0.11.0
numpy==1.20.3
brotlipy==0.7.0
pywin32==227
zipp==3.4.0
lockfile==0.12.2
gensim==3.8.1
~pype1==1.2.1
importlib_metadata==2.0.0
lxml==4.6.1
streamlit==0.82.0
Cython==0.29.21
matplotlib==3.3.2
keyring==21.4.0
pandas==1.2.4
brotli==1.0.9
ipaddr==2.2.0
mypy_extensions==0.4.3
ordereddict==1.1
Pillow==8.2.0
protobuf==3.17.3
pyOpenSSL==20.0.1
wincertstore==0.2.1
